fix issue
If you have text like `(test (test (test)))` with your cursor on the
left-most closing paren, typing `vib` does nothing while `vi)` makes the
region go over the inside of the inner-most parentheses.  This fixes it
so they behave the same.  Specifically, it relaxes the rule that the new
region must properly contain the old region to when there actually is a
region -- IE `(point)` can be outside of the new region.